[fixbug]接收串口数据发生溢出

+++ src/stm32-platform/hal/hal_usart.h
@@ -16,7 +16,7 @@ uint32_t hal_usart_set_stop_bit(enum usart_type_e type, enum usart_stop_bits_e s
 uint32_t hal_usart_set_flow_control(enum usart_type_e type, enum usart_flow_control_e control);
 uint32_t hal_usart_transport_data(enum usart_type_e type, const uint8_t* data, uint32_t length);
 uint32_t hal_usart_receive_data(enum usart_type_e type, uint8_t* data, uint32_t length);
-uint8_t hal_usart_receive_buffer(enum usart_type_e type);
+int32_t hal_usart_receive_buffer(enum usart_type_e type);
 
 uint32_t hal_usart_set_enable(enum usart_type_e type);
